Convert seestar_run.py to use argparse instead of sys.argv
[Argparse](https://docs.python.org/3/library/argparse.html) is a
commonly-used Python library for interpreting command-line input for
scripts. It has a couple of advantages over trying to parse input
manually by pulling from sys.argv - notably, it's less brittle (can
reorganize, rename, add, or remove parameters more easily), it has
built-in input validation, and you can add contextual help messages to
make it easier for users to understand how to use your script.

This diff just sets up argparse in a fairly minimal way, but in the
future, this could be expanded with more options and validation for
parameters.